made some changes to routes
Deleted the ErrorRoute, WebserverBuilder#addErrorHandler now requires a normal Route.
Modified the javadoc for RequestData#getUriParts
Made Webserver use an own Logger Instance
Routes that not exists, are now redirected to the "/" route of the router. (Currently only works for the defaultRouter)

Added a short paragraph about error Handling to the README.md